awesome speed enhancement
near constant rendering time for any pixel size. only making two
ctx.getImageData calls (which are painfully slow), each time one
staggered half of the image is rendered (every other column) which
doesnt overlap when masking.

one small issue is that the custom getImageData function i wrote is bad
with non-int values. as it destroys the offset for the array. right now
the results look good by calling CEIL on the hexel edge and FLOOR on the
apothem when i init the hexel properties. however i think i can skip
that and just call those functions with in the REDNER.getImageData loop
to avoid the rounding errors in other places